test: run Git fixture regression in the integration lane
